requirements.txt
fastapi
uvicorn
streamlit
requests
langchain
langchain-google-genai
langchain-community
tavily-python
python-dotenv